requirements.txt
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
Pillow==10.1.0
Pillow==10.3.0
Requests==2.31.0
scikit_learn==1.3.0
streamlit==1.30.0
